refactor: migrate group/ungroup transactions
