+++ apps/server/src/api-data/rundown/rundown.controller.ts
@@ -3,7 +3,6 @@ import { getErrorMessage } from 'ontime-utils';
 
 import type { Request, Response } from 'express';
 
-import { ungroupEntries, groupEntries } from '../../services/rundown-service/RundownService.js';
 import { getEntryWithId } from '../../services/rundown-service/rundownUtils.js';
 
 import { getCurrentRundown } from './rundown.dao.js';
@@ -39,25 +38,5 @@ export async function rundownGetById(req: Request, res: Response<OntimeEntry | E
   } catch (error) {
     const message = getErrorMessage(error);
     res.status(500).json({ message });
-  }
-}
-
-export async function rundownUngroupEntries(req: Request, res: Response<Rundown | ErrorResponse>) {
-  try {
-    const newRundown = await ungroupEntries(req.params.entryId);
-    res.status(200).send(newRundown);
-  } catch (error) {
-    const message = getErrorMessage(error);
-    res.status(400).send({ message });
-  }
-}
-
-export async function rundownAddToBlock(req: Request, res: Response<Rundown | ErrorResponse>) {
-  try {
-    const newRundown = await groupEntries(req.body.ids);
-    res.status(200).send(newRundown);
-  } catch (error) {
-    const message = getErrorMessage(error);
-    res.status(400).send({ message });
   }
 }
